fixed an error with INIT
basically if a scout could not be placed on the square corresponding to its direction, it broke the EC and the EC stopped making any more towers. This is now fixed. I have just tested this code on circle but it's looking good so far. Time to test on more maps after I take a shower